+++ src/components/ToolBar/DataMenu/SaveWorkspaceToNDExOverwrite.tsx
@@ -16,19 +16,14 @@ import { useCredentialStore } from '../../../store/CredentialStore'
 import { AppConfigContext } from '../../../AppConfigContext'
 import { useMessageStore } from '../../../store/MessageStore'
 import { useWorkspaceStore } from '../../../store/WorkspaceStore'
-import { exportNetworkToCx2 } from '../../../store/io/exportCX'
 import { useNetworkSummaryStore } from '../../../store/NetworkSummaryStore'
 import { useVisualStyleStore } from '../../../store/VisualStyleStore'
 import { useNetworkStore } from '../../../store/NetworkStore'
 import { useTableStore } from '../../../store/TableStore'
 import { useViewModelStore } from '../../../store/ViewModelStore'
-import { Network } from '../../../models/NetworkModel'
-import { IdType } from '../../../models/IdType'
 import { KeycloakContext } from '../../../bootstrap'
 import { useUiStateStore } from '../../../store/UiStateStore'
-import { VisualStyleOptions } from 'src/models/VisualStyleModel/VisualStyleOptions'
-import { useNdexNetwork } from '../../../store/hooks/useNdexNetwork'
-import { NdexNetworkSummary, NetworkView, Table, VisualStyle } from 'src/models'
+import { fetchMyWorkspaces, saveAllNetworks } from '../../../utils/ndex-utils'
 
 export const SaveWorkspaceToNDExOverwriteMenuItem = (
   props: BaseMenuProps,
@@ -74,15 +69,8 @@ export const SaveWorkspaceToNDExOverwriteMenuItem = (
   )
 
   useEffect(() => {
-    const fetchMyWorkspaces = async (): Promise<any> => {
-      const ndexClient = new NDEx(ndexBaseUrl)
-      const token = await getToken()
-      ndexClient.setAuthToken(token)
-      const myWorkspaces = await ndexClient.getUserCyWebWorkspaces()
-      return myWorkspaces
-    }
     if (authenticated) {
-      fetchMyWorkspaces()
+      fetchMyWorkspaces(ndexBaseUrl, getToken)
         .then(function (resultArray) {
           const workspaceIds = resultArray.map(
             (item: { workspaceId: any }) => item.workspaceId,
@@ -96,141 +84,24 @@ export const SaveWorkspaceToNDExOverwriteMenuItem = (
     }
   }, [])
 
-  const saveNetworkToNDEx = async (
-    networkId: string,
-    network: Network,
-    visualStyle: VisualStyle,
-    summary: NdexNetworkSummary,
-    nodeTable: Table,
-    edgeTable: Table,
-    viewModel: NetworkView,
-    visualStyleOptions?: VisualStyleOptions,
-  ): Promise<void> => {
-    const ndexClient = new NDEx(ndexBaseUrl)
-    const accessToken = await getToken()
-    ndexClient.setAuthToken(accessToken)
-    const cx = exportNetworkToCx2(
-      network,
-      visualStyle,
-      summary,
-      nodeTable,
-      edgeTable,
-      visualStyleOptions,
-      viewModel,
-    )
-    await ndexClient.updateNetworkFromRawCX2(networkId, cx)
-    const ndexSummary = await ndexClient.getNetworkSummary(networkId)
-    const newNdexModificationTime = ndexSummary.modificationTime
-    updateSummary(networkId, {
-      modificationTime: newNdexModificationTime,
-    })
-  }
-
-  const saveCopyToNDEx = async (
-    network: Network,
-    visualStyle: VisualStyle,
-    summary: NdexNetworkSummary,
-    nodeTable: Table,
-    edgeTable: Table,
-    viewModel: NetworkView,
-    visualStyleOptions?: VisualStyleOptions,
-  ): Promise<void> => {
-    const ndexClient = new NDEx(ndexBaseUrl)
-    const accessToken = await getToken()
-    ndexClient.setAuthToken(accessToken)
-    const cx = exportNetworkToCx2(
-      network,
-      visualStyle,
-      summary,
-      nodeTable,
-      edgeTable,
-      visualStyleOptions,
-      viewModel,
-      `Copy of ${summary.name}`,
-    )
-    const { uuid } = await ndexClient.createNetworkFromRawCX2(cx)
-    addNetworkToWorkspace(uuid as IdType)
-  }
-
-  const saveAllNetworks = async (): Promise<void> => {
-    for (const networkId of allNetworkId) {
-      let network = networks.get(networkId) as Network
-      let visualStyle = visualStyles[networkId]
-      const summary = summaries[networkId]
-      let nodeTable = tables[networkId]?.nodeTable
-      let edgeTable = tables[networkId]?.edgeTable
-      let networkViews: NetworkView[] = viewModels[networkId]
-      let visualStyleOptions: VisualStyleOptions | undefined =
-        networkVisualStyleOpt[networkId]
-
-      if (!network || !visualStyle || !nodeTable || !edgeTable) {
-        const currentToken = await getToken()
-        const res = await useNdexNetwork(networkId, ndexBaseUrl, currentToken)
-        // Using parentheses to perform destructuring assignment correctly
-        ;({
-          network,
-          nodeTable,
-          edgeTable,
-          visualStyle,
-          networkViews,
-          visualStyleOptions,
-        } = res)
-      }
-      if (summary.isNdex === false) {
-        await saveCopyToNDEx(
-          network,
-          visualStyle,
-          summary,
-          nodeTable,
-          edgeTable,
-          networkViews?.[0],
-          visualStyleOptions,
-        )
-        continue
-      }
-      if (networkModifiedStatus[networkId] === true) {
-        try {
-          await saveNetworkToNDEx(
-            networkId,
-            network,
-            visualStyle,
-            summary,
-            nodeTable,
-            edgeTable,
-            networkViews?.[0],
-            visualStyleOptions,
-          )
-          deleteNetworkModifiedStatus(networkId)
-        } catch (e) {
-          try {
-            await saveCopyToNDEx(
-              network,
-              visualStyle,
-              summary,
-              nodeTable,
-              edgeTable,
-              networkViews?.[0],
-              visualStyleOptions,
-            )
-            addMessage({
-              message: `Unable to save the modified network to NDEx. Instead, saved its copy to NDEx. Error: ${e.message as string}`,
-              duration: 3000,
-            })
-          } catch (e) {
-            addMessage({
-              message: `Unable to save the network or its copy to NDEx. Error: ${e.message as string}`,
-              duration: 3000,
-            })
-            throw e
-          }
-        }
-      }
-    }
-  }
-
   const saveWorkspaceToNDEx = async (): Promise<void> => {
     try {
-      await saveAllNetworks()
+      await saveAllNetworks(
+        getToken,
+        allNetworkId,
+        ndexBaseUrl,
+        addNetworkToWorkspace,
+        networkModifiedStatus,
+        updateSummary,
+        deleteNetworkModifiedStatus,
+        addMessage,
+        networks,
+        visualStyles,
+        summaries,
+        tables,
+        viewModels,
+        networkVisualStyleOpt,
+      )
       const ndexClient = new NDEx(ndexBaseUrl)
       const accessToken = await getToken()
       ndexClient.setAuthToken(accessToken)
